% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_brmsfit.R
\name{make_brmsfit}
\alias{make_brmsfit}
\title{Make \code{brmsfit} Object}
\usage{
make_brmsfit(formula, data, path, ...)
}
\arguments{
\item{formula}{A \code{brms} formula used to generate the checkpoints}

\item{data}{A data frame used to generate the checkpoints}

\item{path}{Character string. The path to the folder, that is used for saving
the checkpoints.}

\item{...}{Additional arguments to be passed to \code{brm}.}
}
\value{
An object of class \code{brmsfit}
}
\description{
This is primarily used internally, wherein the output files of
multiple cmdstanr fits are combined into a single \code{brmsfit} object.
object is converted into a \code{brmsfit} object.
}
\note{
This is primarily an internal function that constructs a \code{brmsfit}
object.
}
