% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkpt_brms.R
\name{print.chkpt_brms}
\alias{print.chkpt_brms}
\title{Print \code{chkpt_brms} Objects}
\usage{
\method{print}{chkpt_brms}(x, ...)
}
\arguments{
\item{x}{Object of class \code{chkpt_brms}}

\item{...}{Currently ignored}
}
\value{
No return value, and used to print the \code{chkpt_brms} object.
}
\description{
Print \code{chkpt_brms} Objects
}
\note{
This function mainly avoids printing out a list,
and it is only used when \code{brmsfit = "FALSE"} in \code{\link[chkptstanr]{chkpt_brms}}.

Typically, after fitting, the posterior draws should be summarized with
\code{\link[chkptstanr]{combine_chkpt_draws}} (assuming \code{brmsfit = "FALSE"}).
}
