% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkpt_setup.R
\name{print.chkpt_setup}
\alias{print.chkpt_setup}
\title{Print \code{chkpt_setup} Object}
\usage{
\method{print}{chkpt_setup}(x, ...)
}
\arguments{
\item{x}{An object of class \code{chkpt_setup}.}

\item{...}{Currently ignored.}
}
\value{
No return value, and used to print the \code{chkpt_setup} object.
}
\description{
Print \code{chkpt_setup} Object
}
\examples{
chkpt_setup <- chkpt_setup(
  iter_sampling = 5000,
  iter_warmup = 2000,
  iter_per_chkpt = 10
)


chkpt_setup
}
