% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkpt_stan.R
\name{print.chkpt_stan}
\alias{print.chkpt_stan}
\title{Print \code{chkpt_stan} Objects}
\usage{
\method{print}{chkpt_stan}(x, ...)
}
\arguments{
\item{x}{Object of class \code{chkpt_stan}}

\item{...}{Currently ignored}
}
\value{
No return value, and used to print the \code{chkpt_stan} object.
}
\description{
Print \code{chkpt_stan} Objects
}
\note{
This function mainly avoids printing out a list.

Typically, after fitting, the posterior draws should be summarized with
\code{\link[chkptstanr]{combine_chkpt_draws}}.
}
