% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkpt-helpers.R
\name{reset_checkpoints}
\alias{reset_checkpoints}
\title{Delete Checkpoint Folders containing samples, keep the model}
\usage{
reset_checkpoints(path, reset = TRUE, recompile = FALSE)
}
\arguments{
\item{path}{(character) The path to the checkpoint folder.}

\item{reset}{(logical) If TRUE, only the checkpoint folders are deleted}

\item{recompile}{(logical) If TRUE, the entire folder is deleted allowing for
a fresh start. If both \code{reset} and \code{recompile} are \code{FALSE},
nothing is done.}
}
\description{
Deletes all checkpoint files and folders under path except for
stan_model/model.stan and stan_model/model.exe. This allows you to restart
the sampling from 0 without recompiling the model.
}
